+++ apps/expo-app/package.json
@@ -15,6 +15,7 @@
     "expo": "~51.0.28",
     "expo-blur": "~13.0.2",
     "expo-constants": "~16.0.2",
+    "expo-image-picker": "^15.0.7",
     "expo-linear-gradient": "^13.0.2",
     "expo-linking": "~6.3.1",
     "expo-location": "~17.0.1",
